requirements.txt
# General
Django==4.2.4
django-environ==0.10.0
Pillow==10.0.0
djangorestframework==3.14.0
django-filter==23.2
drf-nested-routers==0.93.4

# Reformatting
flake8==6.1.0

# Documentation
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1

# Authentication
djangorestframework-simplejwt==5.2.2